now skips over existing files

+++ preprocessing/convert_charades.py
@@ -5,8 +5,8 @@ import torch
 import os
 
 def main():
-    feature_dirs = glob('../data/Charades_v1_features_rgb/*', recursive=True)
-    new_features_dir = '../data/Charades_v1_features_rgb_1024/'
+    feature_dirs = glob('data/Charades_v1_features_rgb/*', recursive=True)
+    new_features_dir = 'data/Charades_v1_features_rgb_1024/'
     if not os.path.exists(new_features_dir):
         os.makedirs(new_features_dir)
 
@@ -18,12 +18,15 @@ def main():
         video_paths = glob(path_cmd, recursive=True)
 
         for path in video_paths:
+            path_array = path.split('/')
+            save_path = new_feature_dir + '/' + path_array[-1][:-4] + '.npy'
+            if os.path.exists(save_path):
+                continue
+
             fc7 = np.loadtxt(path)
             fc7 = torch.Tensor(fc7).cuda()
             map_vid = torch.nn.Linear(4096, 1024).cuda()
             video = map_vid(fc7)
-            path_array = path.split('/')
-            save_path = new_feature_dir + '/' + path_array[-1][:-4] + '.npy'
             with open(save_path, 'wb') as f:
                 np.save(f, video.cpu().detach().numpy())
